Fix: messages_received only collects one message
The test was passing because we get called 5 times and we were looking
for the fifth message.

As soon as you wait for the sixth message, it fails, because the retry
loop only calls the probe five times, and we were only collecting one
message on each iteration.

Instead, collect all available messages on each probe.